Fully Connected Nets layers forward and backward

+++ assignment1/cs231n/features.py
@@ -120,7 +120,7 @@ def hog_feature(im):
         cond2 = temp_ori > 0
         temp_mag = np.where(cond2, grad_mag, 0)
         orientation_histogram[:, :, i] = uniform_filter(temp_mag, size=(cx, cy))[
-            cx / 2::cx, cy / 2::cy].T
+            int(cx / 2)::cx, int(cy / 2)::cy].T
 
     return orientation_histogram.ravel()
